dev-requirements.txt
#
# Please do not manually update this file since the requirements are managed
# by pipenv through Pipfile and Pipfile.lock .
#
# This file is created and managed automatically by the template and it is left
# here only for backwards compatibility reasons with python's ecosystem.
#
# Please use Pipfile to update the requirements.
#
sphinx>=6.2.1
sphinx-rtd-theme>=1.3.0
prospector>=1.9.0
coverage>=7.3.0
nose>=1.3.7
nose-htmloutput>=0.6.0
tox>=4.0.0b2
betamax>=0.8.1
betamax-serializers~=0.2.1
semver>=2.13.0
gitwrapperlib>=1.0.3
twine>=4.0.2
coloredlogs>=15.0.1
emoji>=2.8.0
toml>=0.10.2
boto3-type-annotations-with-docs>=0.3.1